Created a Models and Services Folder and moved files appropriate to those folders
